refactor: backend socket logic
